Create utils folder, move util-test there as tests folder
